+++ ta4j-core/src/main/java/org/ta4j/core/criteria/LinearTransactionCostCriterion.java
@@ -43,7 +43,7 @@ public class LinearTransactionCostCriterion extends AbstractAnalysisCriterion {
     private final double a;
     private final double b;
 
-    private ReturnCriterion grossReturn;
+    private final ReturnCriterion grossReturn;
 
     /**
      * Constructor. (a * x)