+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - **BacktestExecutor** evaluates strategies in parallel when possible
 - **CachedIndicator** synchronize on getValue()
 - **BaseBar** defaults to **`DecimalNum`** type in all constructors
+- add `final` to properties where possible
 
 
 ### Removed/Deprecated